Removing while loop with io_service_.run_one() condition for correct execution
Before this commit, the tcp_interface read method would constantly return a timeout error, even
if data had been read properly. After this commit, the read method returns an OK status if the read
was successful, or TIMEOUT or READ_FAILED depending on the failure type. In the Boost asio library,
the io_service can be run continuously, or run once until an event hander has been dispatched. The
return value of the run_one method was previously used as a while loop exit condition, but this
resulted in the initial behavior describe above, as if the run_one method actually returned after
several even handlers were dispatched, instead of just one. After removing the while loop and using
the method alone, the desired behavior was achieved.