Removing timeout/received flags, adding error checking in timeout handler
Before this commit, the result of a read or timeout was stored in a private variable,
populated by the respective callback. After this commit, the conditional statements that
proviously relied on the flags instead rely on the gloabl error message's value.
Additionally, the timeout handler has added error checking to prevent it from executing fully
when the timer.cancel() method is called after a successful read.