Adding no-timeout option when the timeout_ms argument is 0
Before this commit, a timeout was used in every call TCPInterface::read or
TCPInterface::read_exactly, with a default of 0 ms. After this commit,
the default is set to 0 ms, and if the read or read_exactly methods receive
a 0 timeout request, it will not set a deadline for the timeout, resulting
in a blocking read. This allows for the TCPInterface to behave with a
timeout, or else be used as it was previously.